sess_dataif: squish scan-build warning
clang's scan-build warns we may pass a NULL pointer to unlink(2):
avoid this by only calling unlink if g_status_filename is set.

+++ src/sess_dataif.c
@@ -551,7 +551,7 @@ int main(int argc, char **argv)
 
     run_tunnels(&lo, &ro);
     if (g_status_file) fclose(g_status_file);
-    unlink(g_status_filename);
+    if (g_status_filename) unlink(g_status_filename);
 
     return 0;
 }
